Pathing basics done, basic atmosphere
In PathMangeler there is now a working, very performant (0.04ms for ca 10 nodes distance, scaling mostly with number of obstacles encountered) algorithm to find a way through the the Grid Node graph.
TODO: Renaming of functions to make it usable without wanting to die

Also started fiddeling with graphics settings to make an atmospere. Found out that lights already have a setting that makes them turn off when they are a certain distance away, thats gonna be handy.

For testing rn it spawns a prefab when u click on the ground, just to see shadows/objects